删除错误命名修复feed list
