+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "baseUrl": ".",
+    "baseUrl": "/src/.",
     "inlineSourceMap": true,
     "inlineSources": true,
     "module": "ESNext",
